Issue #8126: Extract reusable code from RequireThisCheck
