+++ config/suppressions.xml
@@ -68,8 +68,6 @@
     <suppress checks="MethodCount" files="[\\/]CommentsIndentationCheck.java$"/>
     <!--VisibilityModifierCheck has 7 options which require 7 additional methods (setters)-->
     <suppress checks="MethodCount" files="[\\/]VisibilityModifierCheck.java$"/>
-    <!--RequireThisCheck has a hierarchy of nested classes which contains a lot of methods. -->
-    <suppress checks="MethodCount" files="[\\/]RequireThisCheck.java$"/>
 
     <!-- we need that set of converters -->
     <suppress checks="ClassDataAbstractionCoupling" files="AutomaticBean\.java"/>
