+++ src/main/java/com/puppycrawl/tools/checkstyle/checks/coding/RequireThisCheck.java
@@ -20,8 +20,6 @@
 package com.puppycrawl.tools.checkstyle.checks.coding;
 
 import java.util.ArrayDeque;
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.Deque;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -29,13 +27,23 @@ import java.util.LinkedList;
 import java.util.Map;
 import java.util.Queue;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 import com.puppycrawl.tools.checkstyle.FileStatefulCheck;
 import com.puppycrawl.tools.checkstyle.api.AbstractCheck;
 import com.puppycrawl.tools.checkstyle.api.DetailAST;
 import com.puppycrawl.tools.checkstyle.api.TokenTypes;
 import com.puppycrawl.tools.checkstyle.utils.CheckUtil;
+import com.puppycrawl.tools.checkstyle.utils.FrameUtil;
+import com.puppycrawl.tools.checkstyle.utils.FrameUtil.AbstractFrame;
+import com.puppycrawl.tools.checkstyle.utils.FrameUtil.AnonymousClassFrame;
+import com.puppycrawl.tools.checkstyle.utils.FrameUtil.BlockFrame;
+import com.puppycrawl.tools.checkstyle.utils.FrameUtil.CatchFrame;
+import com.puppycrawl.tools.checkstyle.utils.FrameUtil.ClassFrame;
+import com.puppycrawl.tools.checkstyle.utils.FrameUtil.ConstructorFrame;
+import com.puppycrawl.tools.checkstyle.utils.FrameUtil.ForFrame;
+import com.puppycrawl.tools.checkstyle.utils.FrameUtil.FrameType;
+import com.puppycrawl.tools.checkstyle.utils.FrameUtil.MemberType;
+import com.puppycrawl.tools.checkstyle.utils.FrameUtil.MethodFrame;
 import com.puppycrawl.tools.checkstyle.utils.ScopeUtil;
 import com.puppycrawl.tools.checkstyle.utils.TokenUtil;
 
@@ -226,47 +234,6 @@ public class RequireThisCheck extends AbstractCheck {
      */
     public static final String MSG_VARIABLE = "require.this.variable";
 
-    /** Set of all declaration tokens. */
-    private static final Set<Integer> DECLARATION_TOKENS = Collections.unmodifiableSet(
-        Arrays.stream(new Integer[] {
-            TokenTypes.VARIABLE_DEF,
-            TokenTypes.CTOR_DEF,
-            TokenTypes.METHOD_DEF,
-            TokenTypes.CLASS_DEF,
-            TokenTypes.ENUM_DEF,
-            TokenTypes.ANNOTATION_DEF,
-            TokenTypes.INTERFACE_DEF,
-            TokenTypes.PARAMETER_DEF,
-            TokenTypes.TYPE_ARGUMENT,
-        }).collect(Collectors.toSet()));
-    /** Set of all assign tokens. */
-    private static final Set<Integer> ASSIGN_TOKENS = Collections.unmodifiableSet(
-        Arrays.stream(new Integer[] {
-            TokenTypes.ASSIGN,
-            TokenTypes.PLUS_ASSIGN,
-            TokenTypes.STAR_ASSIGN,
-            TokenTypes.DIV_ASSIGN,
-            TokenTypes.MOD_ASSIGN,
-            TokenTypes.SR_ASSIGN,
-            TokenTypes.BSR_ASSIGN,
-            TokenTypes.SL_ASSIGN,
-            TokenTypes.BAND_ASSIGN,
-            TokenTypes.BXOR_ASSIGN,
-        }).collect(Collectors.toSet()));
-    /** Set of all compound assign tokens. */
-    private static final Set<Integer> COMPOUND_ASSIGN_TOKENS = Collections.unmodifiableSet(
-        Arrays.stream(new Integer[] {
-            TokenTypes.PLUS_ASSIGN,
-            TokenTypes.STAR_ASSIGN,
-            TokenTypes.DIV_ASSIGN,
-            TokenTypes.MOD_ASSIGN,
-            TokenTypes.SR_ASSIGN,
-            TokenTypes.BSR_ASSIGN,
-            TokenTypes.SL_ASSIGN,
-            TokenTypes.BAND_ASSIGN,
-            TokenTypes.BXOR_ASSIGN,
-        }).collect(Collectors.toSet()));
-
     /** Frame for the currently processed AST. */
     private final Deque<AbstractFrame> current = new ArrayDeque<>();
 
@@ -438,10 +405,11 @@ public class RequireThisCheck extends AbstractCheck {
      * @param frame the class frame where the violation is found.
      */
     private void logViolation(String msgKey, DetailAST ast, AbstractFrame frame) {
-        if (frame.getFrameName().equals(getNearestClassFrameName())) {
+        if (frame.getFrameName().equals(FrameUtil.getNearestClassFrame(current.peek())
+                .getFrameName())) {
             log(ast, msgKey, ast.getText(), "");
         }
-        else if (!(frame instanceof AnonymousClassFrame)) {
+        else if (!(frame instanceof FrameUtil.AnonymousClassFrame)) {
             log(ast, msgKey, ast.getText(), frame.getFrameName() + '.');
         }
     }
@@ -463,9 +431,10 @@ public class RequireThisCheck extends AbstractCheck {
 
         if (!importOrPackage
                 && !typeName
-                && !isDeclarationToken(parentType)
+                && !FrameUtil.isDeclarationToken(parentType)
                 && !isLambdaParameter(ast)) {
-            final AbstractFrame fieldFrame = findClassFrame(ast, false);
+            final AbstractFrame fieldFrame = FrameUtil.findClassFrame(ast, current.peek(),
+                    MemberType.FIELD);
 
             if (fieldFrame != null && ((ClassFrame) fieldFrame).hasInstanceMember(ast)) {
                 frame = getClassFrameWhereViolationIsFound(ast);
@@ -478,14 +447,14 @@ public class RequireThisCheck extends AbstractCheck {
      * Parses the next AST for declarations.
      *
      * @param frameStack stack containing the FrameTree being built.
-     * @param ast AST to parse.
+     * @param ast        AST to parse.
      */
     // -@cs[JavaNCSS] This method is a big switch and is too hard to remove.
     private static void collectDeclarations(Deque<AbstractFrame> frameStack, DetailAST ast) {
         final AbstractFrame frame = frameStack.peek();
         switch (ast.getType()) {
             case TokenTypes.VARIABLE_DEF:
-                collectVariableDeclarations(ast, frame);
+                FrameUtil.collectVariableDeclarations(ast, frame);
                 break;
             case TokenTypes.PARAMETER_DEF:
                 if (!CheckUtil.isReceiverParameter(ast)
@@ -535,7 +504,7 @@ public class RequireThisCheck extends AbstractCheck {
                 frameStack.addFirst(forFrame);
                 break;
             case TokenTypes.LITERAL_NEW:
-                if (isAnonymousClassDef(ast)) {
+                if (CheckUtil.isAnonymousClassDef(ast)) {
                     frameStack.addFirst(new AnonymousClassFrame(frame,
                             ast.getFirstChild().toString()));
                 }
@@ -546,34 +515,10 @@ public class RequireThisCheck extends AbstractCheck {
     }
 
     /**
-     * Collects variable declarations.
-     *
-     * @param ast variable token.
-     * @param frame current frame.
-     */
-    private static void collectVariableDeclarations(DetailAST ast, AbstractFrame frame) {
-        final DetailAST ident = ast.findFirstToken(TokenTypes.IDENT);
-        if (frame.getType() == FrameType.CLASS_FRAME) {
-            final DetailAST mods =
-                    ast.findFirstToken(TokenTypes.MODIFIERS);
-            if (ScopeUtil.isInInterfaceBlock(ast)
-                    || mods.findFirstToken(TokenTypes.LITERAL_STATIC) != null) {
-                ((ClassFrame) frame).addStaticMember(ident);
-            }
-            else {
-                ((ClassFrame) frame).addInstanceMember(ident);
-            }
-        }
-        else {
-            frame.addIdent(ident);
-        }
-    }
-
-    /**
      * Ends parsing of the AST for declarations.
      *
      * @param frameStack Stack containing the FrameTree being built.
-     * @param ast AST that was parsed.
+     * @param ast        AST that was parsed.
      */
     private void endCollectingDeclarations(Queue<AbstractFrame> frameStack, DetailAST ast) {
         switch (ast.getType()) {
@@ -589,25 +534,13 @@ public class RequireThisCheck extends AbstractCheck {
                 frames.put(ast, frameStack.poll());
                 break;
             case TokenTypes.LITERAL_NEW:
-                if (isAnonymousClassDef(ast)) {
+                if (CheckUtil.isAnonymousClassDef(ast)) {
                     frames.put(ast, frameStack.poll());
                 }
                 break;
             default:
                 // do nothing
         }
-    }
-
-    /**
-     * Whether the AST is a definition of an anonymous class.
-     *
-     * @param ast the AST to process.
-     * @return true if the AST is a definition of an anonymous class.
-     */
-    private static boolean isAnonymousClassDef(DetailAST ast) {
-        final DetailAST lastChild = ast.getLastChild();
-        return lastChild != null
-            && lastChild.getType() == TokenTypes.OBJBLOCK;
     }
 
     /**
@@ -621,7 +554,8 @@ public class RequireThisCheck extends AbstractCheck {
     // a logic, additional abstraction will not make logic/algorithm more readable.
     private AbstractFrame getClassFrameWhereViolationIsFound(DetailAST ast) {
         AbstractFrame frameWhereViolationIsFound = null;
-        final AbstractFrame variableDeclarationFrame = findFrame(ast, false);
+        final AbstractFrame variableDeclarationFrame = FrameUtil.findFrame(ast, current.peek(),
+               MemberType.FIELD);
         final FrameType variableDeclarationFrameType = variableDeclarationFrame.getType();
         final DetailAST prevSibling = ast.getPreviousSibling();
         if (variableDeclarationFrameType == FrameType.CLASS_FRAME
@@ -636,22 +570,25 @@ public class RequireThisCheck extends AbstractCheck {
                         && !isReturnedVariable(variableDeclarationFrame, ast)
                         && canBeReferencedFromStaticContext(ast)
                         && canAssignValueToClassField(ast)) {
-                    frameWhereViolationIsFound = findFrame(ast, true);
+                    frameWhereViolationIsFound = FrameUtil.findFrame(ast, current.peek(),
+                            MemberType.METHOD);
                 }
             }
             else if (!validateOnlyOverlapping
                      && prevSibling == null
-                     && isAssignToken(ast.getParent().getType())
+                     && FrameUtil.isAssignToken(ast.getParent().getType())
                      && !isUserDefinedArrangementOfThis(variableDeclarationFrame, ast)
                      && canBeReferencedFromStaticContext(ast)
                      && canAssignValueToClassField(ast)) {
-                frameWhereViolationIsFound = findFrame(ast, true);
+                frameWhereViolationIsFound = FrameUtil.findFrame(ast, current.peek(),
+                        MemberType.METHOD);
             }
         }
         else if (variableDeclarationFrameType == FrameType.CTOR_FRAME
                  && isOverlappingByArgument(ast)
                  && !isUserDefinedArrangementOfThis(variableDeclarationFrame, ast)) {
-            frameWhereViolationIsFound = findFrame(ast, true);
+            frameWhereViolationIsFound = FrameUtil.findFrame(ast, current.peek(),
+                    MemberType.METHOD);
         }
         else if (variableDeclarationFrameType == FrameType.BLOCK_FRAME
                     && isOverlappingByLocalVariable(ast)
@@ -659,7 +596,8 @@ public class RequireThisCheck extends AbstractCheck {
                     && !isUserDefinedArrangementOfThis(variableDeclarationFrame, ast)
                     && !isReturnedVariable(variableDeclarationFrame, ast)
                     && canBeReferencedFromStaticContext(ast)) {
-            frameWhereViolationIsFound = findFrame(ast, true);
+            frameWhereViolationIsFound = FrameUtil.findFrame(ast, current.peek(),
+                    MemberType.METHOD);
         }
         return frameWhereViolationIsFound;
     }
@@ -714,7 +652,8 @@ public class RequireThisCheck extends AbstractCheck {
      * @param blockStartToken token which starts the block.
      * @return the token which ends the code block.
      */
-    private static DetailAST getBlockEndToken(DetailAST blockNameIdent, DetailAST blockStartToken) {
+    private static DetailAST getBlockEndToken(DetailAST blockNameIdent,
+                                              DetailAST blockStartToken) {
         DetailAST blockEndToken = null;
         final DetailAST blockNameIdentParent = blockNameIdent.getParent();
         if (blockNameIdentParent.getType() == TokenTypes.CASE_GROUP) {
@@ -769,7 +708,7 @@ public class RequireThisCheck extends AbstractCheck {
     private static boolean isAstInside(DetailAST tree, DetailAST ast) {
         boolean result = false;
 
-        if (isAstSimilar(tree, ast)) {
+        if (TokenUtil.isAstSimilar(tree, ast)) {
             result = true;
         }
         else {
@@ -789,7 +728,8 @@ public class RequireThisCheck extends AbstractCheck {
      * @return true if field can be referenced from a static context.
      */
     private boolean canBeReferencedFromStaticContext(DetailAST ident) {
-        AbstractFrame variableDeclarationFrame = findFrame(ident, false);
+        AbstractFrame variableDeclarationFrame = FrameUtil.findFrame(ident, current.peek(),
+                MemberType.FIELD);
         boolean staticInitializationBlock = false;
         while (variableDeclarationFrame.getType() == FrameType.BLOCK_FRAME
                 || variableDeclarationFrame.getType() == FrameType.FOR_FRAME) {
@@ -852,10 +792,12 @@ public class RequireThisCheck extends AbstractCheck {
      * @return true if a value can be assigned to a field.
      */
     private boolean canAssignValueToClassField(DetailAST ast) {
-        final AbstractFrame fieldUsageFrame = findFrame(ast, false);
+        final AbstractFrame fieldUsageFrame = FrameUtil.findFrame(ast, current.peek(),
+                MemberType.FIELD);
         final boolean fieldUsageInConstructor = isInsideConstructorFrame(fieldUsageFrame);
 
-        final AbstractFrame declarationFrame = findFrame(ast, true);
+        final AbstractFrame declarationFrame = FrameUtil.findFrame(ast, current.peek(),
+                MemberType.METHOD);
         final boolean finalField = ((ClassFrame) declarationFrame).hasFinalField(ast);
 
         return fieldUsageInConstructor || !finalField;
@@ -891,12 +833,13 @@ public class RequireThisCheck extends AbstractCheck {
         boolean overlapping = false;
         final DetailAST parent = ast.getParent();
         final DetailAST sibling = ast.getNextSibling();
-        if (sibling != null && isAssignToken(parent.getType())) {
-            if (isCompoundAssignToken(parent.getType())) {
+        if (sibling != null && FrameUtil.isAssignToken(parent.getType())) {
+            if (FrameUtil.isCompoundAssignToken(parent.getType())) {
                 overlapping = true;
             }
             else {
-                final ClassFrame classFrame = (ClassFrame) findFrame(ast, true);
+                final ClassFrame classFrame = (ClassFrame) FrameUtil.findFrame(ast,
+                        current.peek(), MemberType.METHOD);
                 final Set<DetailAST> exprIdents = getAllTokensOfType(sibling, TokenTypes.IDENT);
                 overlapping = classFrame.containsFieldOrVariableDef(exprIdents, ast);
             }
@@ -914,8 +857,9 @@ public class RequireThisCheck extends AbstractCheck {
         boolean overlapping = false;
         final DetailAST parent = ast.getParent();
         final DetailAST sibling = ast.getNextSibling();
-        if (sibling != null && isAssignToken(parent.getType())) {
-            final ClassFrame classFrame = (ClassFrame) findFrame(ast, true);
+        if (sibling != null && FrameUtil.isAssignToken(parent.getType())) {
+            final ClassFrame classFrame = (ClassFrame) FrameUtil.findFrame(ast, current.peek(),
+                    MemberType.METHOD);
             final Set<DetailAST> exprIdents = getAllTokensOfType(sibling, TokenTypes.IDENT);
             overlapping = classFrame.containsFieldOrVariableDef(exprIdents, ast);
         }
@@ -993,7 +937,8 @@ public class RequireThisCheck extends AbstractCheck {
      * @return a set of tokens which are equal to current token starting with the current ast node
      *         and which line number is lower or equal to the end line number.
      */
-    private static Set<DetailAST> getAllTokensWhichAreEqualToCurrent(DetailAST ast, DetailAST token,
+    private static Set<DetailAST> getAllTokensWhichAreEqualToCurrent(DetailAST ast,
+                                                                     DetailAST token,
                                                                      int endLineNumber) {
         DetailAST vertex = ast;
         final Set<DetailAST> result = new HashSet<>();
@@ -1003,7 +948,7 @@ public class RequireThisCheck extends AbstractCheck {
                 vertex = stack.pop();
             }
             while (vertex != null) {
-                if (isAstSimilar(token, vertex)
+                if (TokenUtil.isAstSimilar(token, vertex)
                         && vertex.getLineNo() <= endLineNumber) {
                     result.add(vertex);
                 }
@@ -1027,7 +972,8 @@ public class RequireThisCheck extends AbstractCheck {
     private AbstractFrame getMethodWithoutThis(DetailAST ast) {
         AbstractFrame result = null;
         if (!validateOnlyOverlapping) {
-            final AbstractFrame frame = findFrame(ast, true);
+            final AbstractFrame frame = FrameUtil.findFrame(ast, current.peek(),
+                    MemberType.METHOD);
             if (frame != null
                     && ((ClassFrame) frame).hasInstanceMethod(ast)
                     && !((ClassFrame) frame).hasStaticMethod(ast)) {
@@ -1035,96 +981,6 @@ public class RequireThisCheck extends AbstractCheck {
             }
         }
         return result;
-    }
-
-    /**
-     * Find the class frame containing declaration.
-     *
-     * @param name IDENT ast of the declaration to find.
-     * @param lookForMethod whether we are looking for a method name.
-     * @return AbstractFrame containing declaration or null.
-     */
-    private AbstractFrame findClassFrame(DetailAST name, boolean lookForMethod) {
-        AbstractFrame frame = current.peek();
-
-        while (true) {
-            frame = findFrame(frame, name, lookForMethod);
-
-            if (frame == null || frame instanceof ClassFrame) {
-                break;
-            }
-
-            frame = frame.getParent();
-        }
-
-        return frame;
-    }
-
-    /**
-     * Find frame containing declaration.
-     *
-     * @param name IDENT ast of the declaration to find.
-     * @param lookForMethod whether we are looking for a method name.
-     * @return AbstractFrame containing declaration or null.
-     */
-    private AbstractFrame findFrame(DetailAST name, boolean lookForMethod) {
-        return findFrame(current.peek(), name, lookForMethod);
-    }
-
-    /**
-     * Find frame containing declaration.
-     *
-     * @param frame The parent frame to searching in.
-     * @param name IDENT ast of the declaration to find.
-     * @param lookForMethod whether we are looking for a method name.
-     * @return AbstractFrame containing declaration or null.
-     */
-    private static AbstractFrame findFrame(AbstractFrame frame, DetailAST name,
-            boolean lookForMethod) {
-        return frame.getIfContains(name, lookForMethod);
-    }
-
-    /**
-     * Check that token is related to Definition tokens.
-     *
-     * @param parentType token Type.
-     * @return true if token is related to Definition Tokens.
-     */
-    private static boolean isDeclarationToken(int parentType) {
-        return DECLARATION_TOKENS.contains(parentType);
-    }
-
-    /**
-     * Check that token is related to assign tokens.
-     *
-     * @param tokenType token type.
-     * @return true if token is related to assign tokens.
-     */
-    private static boolean isAssignToken(int tokenType) {
-        return ASSIGN_TOKENS.contains(tokenType);
-    }
-
-    /**
-     * Check that token is related to compound assign tokens.
-     *
-     * @param tokenType token type.
-     * @return true if token is related to compound assign tokens.
-     */
-    private static boolean isCompoundAssignToken(int tokenType) {
-        return COMPOUND_ASSIGN_TOKENS.contains(tokenType);
-    }
-
-    /**
-     * Gets the name of the nearest parent ClassFrame.
-     *
-     * @return the name of the nearest parent ClassFrame.
-     */
-    private String getNearestClassFrameName() {
-        AbstractFrame frame = current.peek();
-        while (frame.getType() != FrameType.CLASS_FRAME) {
-            frame = frame.getParent();
-        }
-        return frame.getFrameName();
     }
 
     /**
@@ -1162,488 +1018,4 @@ public class RequireThisCheck extends AbstractCheck {
         }
         return isLambdaParameter;
     }
-
-    /**
-     * Checks if 2 AST are similar by their type and text.
-     *
-     * @param left The first AST to check.
-     * @param right The second AST to check.
-     * @return {@code true} if they are similar.
-     */
-    private static boolean isAstSimilar(DetailAST left, DetailAST right) {
-        return left.getType() == right.getType() && left.getText().equals(right.getText());
-    }
-
-    /** An AbstractFrame type. */
-    private enum FrameType {
-
-        /** Class frame type. */
-        CLASS_FRAME,
-        /** Constructor frame type. */
-        CTOR_FRAME,
-        /** Method frame type. */
-        METHOD_FRAME,
-        /** Block frame type. */
-        BLOCK_FRAME,
-        /** Catch frame type. */
-        CATCH_FRAME,
-        /** For frame type. */
-        FOR_FRAME,
-
-    }
-
-    /**
-     * A declaration frame.
-     */
-    private abstract static class AbstractFrame {
-
-        /** Set of name of variables declared in this frame. */
-        private final Set<DetailAST> varIdents;
-
-        /** Parent frame. */
-        private final AbstractFrame parent;
-
-        /** Name identifier token. */
-        private final DetailAST frameNameIdent;
-
-        /**
-         * Constructor -- invocable only via super() from subclasses.
-         *
-         * @param parent parent frame.
-         * @param ident frame name ident.
-         */
-        protected AbstractFrame(AbstractFrame parent, DetailAST ident) {
-            this.parent = parent;
-            frameNameIdent = ident;
-            varIdents = new HashSet<>();
-        }
-
-        /**
-         * Get the type of the frame.
-         *
-         * @return a FrameType.
-         */
-        protected abstract FrameType getType();
-
-        /**
-         * Add a name to the frame.
-         *
-         * @param identToAdd the name we're adding.
-         */
-        private void addIdent(DetailAST identToAdd) {
-            varIdents.add(identToAdd);
-        }
-
-        protected AbstractFrame getParent() {
-            return parent;
-        }
-
-        protected String getFrameName() {
-            return frameNameIdent.getText();
-        }
-
-        public DetailAST getFrameNameIdent() {
-            return frameNameIdent;
-        }
-
-        /**
-         * Check whether the frame contains a field or a variable with the given name.
-         *
-         * @param nameToFind the IDENT ast of the name we're looking for.
-         * @return whether it was found.
-         */
-        protected boolean containsFieldOrVariable(DetailAST nameToFind) {
-            return containsFieldOrVariableDef(varIdents, nameToFind);
-        }
-
-        /**
-         * Check whether the frame contains a given name.
-         *
-         * @param nameToFind IDENT ast of the name we're looking for.
-         * @param lookForMethod whether we are looking for a method name.
-         * @return whether it was found.
-         */
-        protected AbstractFrame getIfContains(DetailAST nameToFind, boolean lookForMethod) {
-            final AbstractFrame frame;
-
-            if (!lookForMethod
-                && containsFieldOrVariable(nameToFind)) {
-                frame = this;
-            }
-            else {
-                frame = parent.getIfContains(nameToFind, lookForMethod);
-            }
-            return frame;
-        }
-
-        /**
-         * Whether the set contains a declaration with the text of the specified
-         * IDENT ast and it is declared in a proper position.
-         *
-         * @param set the set of declarations.
-         * @param ident the specified IDENT ast.
-         * @return true if the set contains a declaration with the text of the specified
-         *         IDENT ast and it is declared in a proper position.
-         */
-        protected boolean containsFieldOrVariableDef(Set<DetailAST> set, DetailAST ident) {
-            boolean result = false;
-            for (DetailAST ast: set) {
-                if (isProperDefinition(ident, ast)) {
-                    result = true;
-                    break;
-                }
-            }
-            return result;
-        }
-
-        /**
-         * Whether the definition is correspondent to the IDENT.
-         *
-         * @param ident the IDENT ast to check.
-         * @param ast the IDENT ast of the definition to check.
-         * @return true if ast is correspondent to ident.
-         */
-        protected boolean isProperDefinition(DetailAST ident, DetailAST ast) {
-            final String nameToFind = ident.getText();
-            return nameToFind.equals(ast.getText())
-                && CheckUtil.isBeforeInSource(ast, ident);
-        }
-    }
-
-    /**
-     * A frame initiated at method definition; holds a method definition token.
-     */
-    private static class MethodFrame extends AbstractFrame {
-
-        /**
-         * Creates method frame.
-         *
-         * @param parent parent frame.
-         * @param ident method name identifier token.
-         */
-        protected MethodFrame(AbstractFrame parent, DetailAST ident) {
-            super(parent, ident);
-        }
-
-        @Override
-        protected FrameType getType() {
-            return FrameType.METHOD_FRAME;
-        }
-
-    }
-
-    /**
-     * A frame initiated at constructor definition.
-     */
-    private static class ConstructorFrame extends AbstractFrame {
-
-        /**
-         * Creates a constructor frame.
-         *
-         * @param parent parent frame.
-         * @param ident frame name ident.
-         */
-        protected ConstructorFrame(AbstractFrame parent, DetailAST ident) {
-            super(parent, ident);
-        }
-
-        @Override
-        protected FrameType getType() {
-            return FrameType.CTOR_FRAME;
-        }
-
-    }
-
-    /**
-     * A frame initiated at class, enum or interface definition; holds instance variable names.
-     */
-    private static class ClassFrame extends AbstractFrame {
-
-        /** Set of idents of instance members declared in this frame. */
-        private final Set<DetailAST> instanceMembers;
-        /** Set of idents of instance methods declared in this frame. */
-        private final Set<DetailAST> instanceMethods;
-        /** Set of idents of variables declared in this frame. */
-        private final Set<DetailAST> staticMembers;
-        /** Set of idents of static methods declared in this frame. */
-        private final Set<DetailAST> staticMethods;
-
-        /**
-         * Creates new instance of ClassFrame.
-         *
-         * @param parent parent frame.
-         * @param ident frame name ident.
-         */
-        /* package */ ClassFrame(AbstractFrame parent, DetailAST ident) {
-            super(parent, ident);
-            instanceMembers = new HashSet<>();
-            instanceMethods = new HashSet<>();
-            staticMembers = new HashSet<>();
-            staticMethods = new HashSet<>();
-        }
-
-        @Override
-        protected FrameType getType() {
-            return FrameType.CLASS_FRAME;
-        }
-
-        /**
-         * Adds static member's ident.
-         *
-         * @param ident an ident of static member of the class.
-         */
-        public void addStaticMember(final DetailAST ident) {
-            staticMembers.add(ident);
-        }
-
-        /**
-         * Adds static method's name.
-         *
-         * @param ident an ident of static method of the class.
-         */
-        public void addStaticMethod(final DetailAST ident) {
-            staticMethods.add(ident);
-        }
-
-        /**
-         * Adds instance member's ident.
-         *
-         * @param ident an ident of instance member of the class.
-         */
-        public void addInstanceMember(final DetailAST ident) {
-            instanceMembers.add(ident);
-        }
-
-        /**
-         * Adds instance method's name.
-         *
-         * @param ident an ident of instance method of the class.
-         */
-        public void addInstanceMethod(final DetailAST ident) {
-            instanceMethods.add(ident);
-        }
-
-        /**
-         * Checks if a given name is a known instance member of the class.
-         *
-         * @param ident the IDENT ast of the name to check.
-         * @return true is the given name is a name of a known
-         *         instance member of the class.
-         */
-        public boolean hasInstanceMember(final DetailAST ident) {
-            return containsFieldOrVariableDef(instanceMembers, ident);
-        }
-
-        /**
-         * Checks if a given name is a known instance method of the class.
-         *
-         * @param ident the IDENT ast of the method call to check.
-         * @return true if the given ast is correspondent to a known
-         *         instance method of the class.
-         */
-        public boolean hasInstanceMethod(final DetailAST ident) {
-            return containsMethodDef(instanceMethods, ident);
-        }
-
-        /**
-         * Checks if a given name is a known static method of the class.
-         *
-         * @param ident the IDENT ast of the method call to check.
-         * @return true is the given ast is correspondent to a known
-         *         instance method of the class.
-         */
-        public boolean hasStaticMethod(final DetailAST ident) {
-            return containsMethodDef(staticMethods, ident);
-        }
-
-        /**
-         * Checks whether given instance member has final modifier.
-         *
-         * @param instanceMember an instance member of a class.
-         * @return true if given instance member has final modifier.
-         */
-        public boolean hasFinalField(final DetailAST instanceMember) {
-            boolean result = false;
-            for (DetailAST member : instanceMembers) {
-                final DetailAST mods = member.getParent().findFirstToken(TokenTypes.MODIFIERS);
-                final boolean finalMod = mods.findFirstToken(TokenTypes.FINAL) != null;
-                if (finalMod && isAstSimilar(member, instanceMember)) {
-                    result = true;
-                    break;
-                }
-            }
-            return result;
-        }
-
-        @Override
-        protected boolean containsFieldOrVariable(DetailAST nameToFind) {
-            return containsFieldOrVariableDef(instanceMembers, nameToFind)
-                    || containsFieldOrVariableDef(staticMembers, nameToFind);
-        }
-
-        @Override
-        protected boolean isProperDefinition(DetailAST ident, DetailAST ast) {
-            final String nameToFind = ident.getText();
-            return nameToFind.equals(ast.getText());
-        }
-
-        @Override
-        protected AbstractFrame getIfContains(DetailAST nameToFind, boolean lookForMethod) {
-            AbstractFrame frame = null;
-
-            if (lookForMethod && containsMethod(nameToFind)
-                || containsFieldOrVariable(nameToFind)) {
-                frame = this;
-            }
-            else if (getParent() != null) {
-                frame = getParent().getIfContains(nameToFind, lookForMethod);
-            }
-            return frame;
-        }
-
-        /**
-         * Check whether the frame contains a given method.
-         *
-         * @param methodToFind the AST of the method to find.
-         * @return true, if a method with the same name and number of parameters is found.
-         */
-        private boolean containsMethod(DetailAST methodToFind) {
-            return containsMethodDef(instanceMethods, methodToFind)
-                || containsMethodDef(staticMethods, methodToFind);
-        }
-
-        /**
-         * Whether the set contains a method definition with the
-         *     same name and number of parameters.
-         *
-         * @param set the set of definitions.
-         * @param ident the specified method call IDENT ast.
-         * @return true if the set contains a definition with the
-         *     same name and number of parameters.
-         */
-        private static boolean containsMethodDef(Set<DetailAST> set, DetailAST ident) {
-            boolean result = false;
-            for (DetailAST ast: set) {
-                if (isSimilarSignature(ident, ast)) {
-                    result = true;
-                    break;
-                }
-            }
-            return result;
-        }
-
-        /**
-         * Whether the method definition has the same name and number of parameters.
-         *
-         * @param ident the specified method call IDENT ast.
-         * @param ast the ast of a method definition to compare with.
-         * @return true if a method definition has the same name and number of parameters
-         *     as the method call.
-         */
-        private static boolean isSimilarSignature(DetailAST ident, DetailAST ast) {
-            boolean result = false;
-            final DetailAST elistToken = ident.getParent().findFirstToken(TokenTypes.ELIST);
-            if (elistToken != null && ident.getText().equals(ast.getText())) {
-                final int paramsNumber =
-                    ast.getParent().findFirstToken(TokenTypes.PARAMETERS).getChildCount();
-                final int argsNumber = elistToken.getChildCount();
-                result = paramsNumber == argsNumber;
-            }
-            return result;
-        }
-
-    }
-
-    /**
-     * An anonymous class frame; holds instance variable names.
-     */
-    private static class AnonymousClassFrame extends ClassFrame {
-
-        /** The name of the frame. */
-        private final String frameName;
-
-        /**
-         * Creates anonymous class frame.
-         *
-         * @param parent parent frame.
-         * @param frameName name of the frame.
-         */
-        protected AnonymousClassFrame(AbstractFrame parent, String frameName) {
-            super(parent, null);
-            this.frameName = frameName;
-        }
-
-        @Override
-        protected String getFrameName() {
-            return frameName;
-        }
-
-    }
-
-    /**
-     * A frame initiated on entering a statement list; holds local variable names.
-     */
-    private static class BlockFrame extends AbstractFrame {
-
-        /**
-         * Creates block frame.
-         *
-         * @param parent parent frame.
-         * @param ident ident frame name ident.
-         */
-        protected BlockFrame(AbstractFrame parent, DetailAST ident) {
-            super(parent, ident);
-        }
-
-        @Override
-        protected FrameType getType() {
-            return FrameType.BLOCK_FRAME;
-        }
-
-    }
-
-    /**
-     * A frame initiated on entering a catch block; holds local catch variable names.
-     */
-    private static class CatchFrame extends AbstractFrame {
-
-        /**
-         * Creates catch frame.
-         *
-         * @param parent parent frame.
-         * @param ident ident frame name ident.
-         */
-        protected CatchFrame(AbstractFrame parent, DetailAST ident) {
-            super(parent, ident);
-        }
-
-        @Override
-        public FrameType getType() {
-            return FrameType.CATCH_FRAME;
-        }
-
-    }
-
-    /**
-     * A frame initiated on entering a for block; holds local for variable names.
-     */
-    private static class ForFrame extends AbstractFrame {
-
-        /**
-         * Creates for frame.
-         *
-         * @param parent parent frame.
-         * @param ident ident frame name ident.
-         */
-        protected ForFrame(AbstractFrame parent, DetailAST ident) {
-            super(parent, ident);
-        }
-
-        @Override
-        public FrameType getType() {
-            return FrameType.FOR_FRAME;
-        }
-
-    }
-
 }
